+++ CHANGELOG.md
@@ -38,6 +38,7 @@ v5.5.0
 - Add an exact solution for GetLocCoords of straight-edge quad elements (!1704)
 - Remove MaxIterations parameter from AssemblyMap (!1710)
 - Consistently use relative tolerance for GMRES (!1706)
+- Fix use of absolute tolerance for iterative solvers (!1711)
 
 **ADRSolver**
 - Add support for spatially-constant, but variable direction, diffusion to